requirements.txt
Django==1.9.6
djangorestframework==3.3.3
numpy==1.11.0
pandas==0.18.0
psycopg2==2.6.1
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
